+++ projects/igniteui-angular/core/src/services/overlay/position/container-position-strategy.ts
@@ -1,4 +1,4 @@
-import { PositionSettings, Util } from '../utilities';
+import { Point, PositionSettings, Size, Util } from '../utilities';
 import { GlobalPositionStrategy } from './global-position-strategy';
 
 /**
@@ -14,16 +14,20 @@ export class ContainerPositionStrategy extends GlobalPositionStrategy {
     /**
      * Position the element based on the PositionStrategy implementing this interface.
      */
-    public override position(contentElement: HTMLElement): void {
+    public override position(contentElement: HTMLElement, _size?: Size, _document?: Document, _initialCall?: boolean, target?: Point | HTMLElement): void {
         // Set up intersection observer
         this.io?.disconnect();
         const outletElement = contentElement.parentElement.parentElement;
+        if (!target || target instanceof Point || !outletElement) {
+            super.position(contentElement, _size, _document, _initialCall, target);
+            return;
+        }
         this.io = Util.setupIntersectionObserver(
-            outletElement,
+            target as HTMLElement || outletElement,
             contentElement.ownerDocument,
-            () => this.updatePosition(contentElement)
+            () => this.updatePosition(contentElement, target as HTMLElement)
         );
-        this.internalPosition(contentElement);
+        this.internalPosition(contentElement, target as HTMLElement);
     }
 
     /**
@@ -34,23 +38,33 @@ export class ContainerPositionStrategy extends GlobalPositionStrategy {
         this.io = null;
     }
 
-    private internalPosition(contentElement: HTMLElement): void {
+    private internalPosition(contentElement: HTMLElement, targetElement: HTMLElement): void {
         contentElement.classList.add('igx-overlay__content--relative');
         contentElement.parentElement.classList.add('igx-overlay__wrapper--flex-container');
         this.setPosition(contentElement);
-        this.updatePosition(contentElement);
+        this.updatePosition(contentElement, targetElement);
     }
 
-    private updatePosition(contentElement: HTMLElement): void {
+    private updatePosition(contentElement: HTMLElement, targetElement: HTMLElement): void {
         const outletElement = contentElement.parentElement?.parentElement;
-        if (!outletElement)
+        if (!targetElement && !outletElement)
             return;
 
-        // TODO: consider using new anchor() CSS function when it becomes more widely supported: https://caniuse.com/mdn-css_properties_anchor
-        const outletRect = outletElement.getBoundingClientRect();
-        contentElement.parentElement.style.width = `${outletRect.width}px`;
-        contentElement.parentElement.style.height = `${outletRect.height}px`;
-        contentElement.parentElement.style.top = `${outletRect.top}px`;
-        contentElement.parentElement.style.left = `${outletRect.left}px`;
+        // TODO: consider using new anchor() CSS function when it becomes more widely
+        // supported: https://caniuse.com/mdn-css_properties_anchor
+        if (outletElement) {
+            const outletRect = outletElement.getBoundingClientRect();
+            contentElement.parentElement.style.width = `${outletRect.width}px`;
+            contentElement.parentElement.style.height = `${outletRect.height}px`;
+            contentElement.parentElement.style.top = `${outletRect.top}px`;
+            contentElement.parentElement.style.left = `${outletRect.left}px`;
+        }
+        if (targetElement) {
+            const targetRect = targetElement.getBoundingClientRect();
+            contentElement.parentElement.style.width = `${targetRect.width}px`;
+            contentElement.parentElement.style.height = `${targetRect.height}px`;
+            contentElement.parentElement.style.top = `${targetRect.top}px`;
+            contentElement.parentElement.style.left = `${targetRect.left}px`;
+        }
     }
 }