+++ projects/igniteui-angular/date-picker/src/date-picker/calendar-container/calendar-container.component.ts
@@ -64,6 +64,7 @@ export class IgxCalendarContainerComponent {
     @HostListener('keydown.alt.arrowup', ['$event'])
     public onEscape(event) {
         event.preventDefault();
+        event.stopPropagation();
         this.calendarClose.emit();
     }